requirements.txt
streamlit
tensorflow
opencv-python
mediapipe==0.10.9
protobuf==3.20.3
pillow
numpy
matplotlib